requirements.txt
fastapi
uvicorn
requests
pygsheets
python-multipart
python-dotenv
sqlalchemy
boto3
supabase
apscheduler
imap-tools
clerk-sdk-python 